add disabled when submitting to prevent multiple database insertion

+++ database/migrations/2016_09_19_162240_create_settings_table.php
@@ -18,7 +18,7 @@ class CreateSettingsTable extends Migration
             $table->string('name_en')->nullable();
             $table->string('name_th');
             $table->string('code')->nullable();
-            $table->float('value')->default(0);
+            $table->double('value')->default(0);
             $table->string('unit_of_measure')->nullable();
 
             $table->integer('group_id')->unsigned()->nullable();
